requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
pyTelegramBotAPI==4.5.1
requests==2.27.1
urllib3==1.26.9
python_dotenv=1.0.1
psycopg2=2.9.10